--- a/lib/logasm.rb
+++ b/lib/logasm.rb
@@ -1,4 +1,3 @@
-require 'inflecto'
 require 'logger'
 require 'oj'
 require_relative 'logasm/adapters'