--- a/lib/logasm/preprocessors.rb
+++ b/lib/logasm/preprocessors.rb
@@ -1,8 +1,17 @@
 class Logasm
   module Preprocessors
     def self.get(type, arguments)
-      require_relative "preprocessors/#{type.to_s}"
-      preprocessor = const_get(Inflecto.camelize(type.to_s))
+      preprocessor =
+        case type.to_s
+        when 'blacklist'
+          require_relative 'preprocessors/blacklist'
+          Preprocessors::Blacklist
+        when 'whitelist'
+          require_relative 'preprocessors/whitelist'
+          Preprocessors::Whitelist
+        else
+          raise "Unknown preprocessor: #{type}"
+        end
       preprocessor.new(arguments)
     end
   end