--- a/logasm.gemspec
+++ b/logasm.gemspec
@@ -16,7 +16,6 @@ Gem::Specification.new do |gem|
   gem.test_files    = gem.files.grep(%r{^(test|spec|features)/})
   gem.require_paths = ["lib"]
 
-  gem.add_dependency 'inflecto'
   gem.add_dependency 'lru_redux'
   gem.add_dependency 'oj'