--- a/spec/utils_spec.rb
+++ b/spec/utils_spec.rb
@@ -2,28 +2,18 @@ require 'spec_helper'
 
 describe Logasm::Utils do
   describe '.build_event' do
-    subject(:event) { described_class.build_event(metadata, level, service_name) }
+    subject(:event) { described_class.build_event(metadata, level, application_name) }
 
-    let(:service_name) { 'test_service' }
-    let(:level)        { 'INFO' }
-    let(:metadata)     { {x: 'y'} }
+    let(:application_name) { 'test_service' }
+    let(:level)  { 'INFO' }
+    let(:metadata) { {x: 'y'} }
 
     before do
       allow(Time).to receive(:now) { Time.utc(2015, 10, 11, 23, 10, 21, 123456) }
     end
 
-    context 'when service name is in correct format' do
-      it 'includes it in the event as application' do
-        expect(event[:application]).to eq('test_service')
-      end
-    end
-
-    context 'when service name is in camelcase' do
-      let(:service_name) { 'InformationService' }
-
-      it 'includes it in the event as lower snake case' do
-        expect(event[:application]).to eq('information_service')
-      end
+    it 'includes it in the event as application' do
+      expect(event[:application]).to eq(application_name)
     end
 
     it 'includes level as a lower case string' do