Tested performance using: `time ruby perf.rb > /dev/null`

where perf.rb was:
```
logger = Logasm.build('test_service', stdout: {level: 'info', json: true})
100_000.times do
  logger.info "hello there", a: "asdf", b: 'eadsfasdf', c: {hello: 'there'}
end
```

Previously: `3.29s user 0.06s system 99% cpu 3.354 total`
Now: `2.07s user 0.04s system 99% cpu 2.117 total`

The tests were done on my laptop.